@@ -1,7 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import StaffingSystem from './pages/formpage';
-ReactDOM.render(
-    <StaffingSystem />,
-  document.getElementById('root')
-);
+function App () {
+  return <StaffingSystem />
+}
+export default App
